requirements.txt
aiofiles==25.1.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.17.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
Authlib==1.6.5
bcrypt==5.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.2
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
curl_cffi==0.11.3
cycler==0.12.1
ecdsa==0.19.1
extra-streamlit-components==0.1.81
fastapi==0.104.1
filelock==3.20.0
fonttools==4.58.2
frozendict==2.4.6
fsspec==2025.10.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.43.0
google-generativeai==0.3.2
googleapis-common-protos==1.72.0
googlemaps==4.10.0
greenlet==3.2.3
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
langchain==1.0.7
langchain-core==1.0.5
langchain-google-genai==0.0.1
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
lxml==5.4.0
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
missingno==0.5.2
mpmath==1.3.0
multitasking==0.0.11
mysql-connector-python==9.5.0
narwhals==2.12.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.0
pandas-datareader==0.10.0
passlib==1.7.4
patsy==1.0.1
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==9.0.1
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
requests-toolbelt==1.0.0
researchpy==0.3.6
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
setuptools==80.9.0
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
SKompiler==0.7
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.27.0
statsmodels==0.14.4
streamlit==1.51.0
stringzilla==4.2.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.22
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.24.0
watchdog==6.0.0
websockets==15.0.1
xxhash==3.6.0
yfinance==0.2.62
zstandard==0.25.0